# Tekst-API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tekst-Dev",
    "description": "An online text research platform",
    "termsOfService": "None",
    "contact": {},
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
    "/admin/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get statistics",
        "operationId": "getStatistics",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformStats"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get users",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "type": "array",
                  "title": "Response Get Users Admin Users Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LocaleKey": {
        "type": "string",
        "enum": [
          "deDE",
          "enUS"
        ]
      },
...
      "PlatformStats": {
        "properties": {
          "usersCount": {
            "type": "integer",
            "title": "Userscount"
          },
          "texts": {
            "items": {
              "$ref": "#/components/schemas/TextStats"
            },
            "type": "array",
            "title": "Texts"
          }
        },
        "type": "object",
        "required": [
          "usersCount",
          "texts"
        ],
        "title": "PlatformStats",
        "description": "Platform statistics data"
      },
...
      "TextStats": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "locationsCount": {
            "type": "integer",
            "title": "Locationscount"
          },
          "resourcesCount": {
            "type": "integer",
            "title": "Resourcescount"
          },
          "resourceTypes": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Resourcetypes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "locationsCount",
          "resourcesCount",
          "resourceTypes"
        ],
        "title": "TextStats",
        "description": "Text statistics data"
      },
...
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "isSuperuser": {
            "type": "boolean",
            "title": "Issuperuser"
          },
          "isVerified": {
            "type": "boolean",
            "title": "Isverified"
          },
          "username": {
            "type": "string",
            "maxLength": 16,
            "minLength": 4,
            "pattern": "[a-zA-Z0-9\\-_]+",
            "title": "Username"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          },
          "affiliation": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Affiliation"
          },
          "locale": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocaleKey"
              },
              {
                "type": "null"
              }
            ]
          },
          "publicFields": {
            "items": {
              "type": "string",
              "enum": [
                "name",
                "affiliation"
              ]
            },
            "type": "array",
            "title": "Publicfields",
            "description": "Data fields set public by this user",
            "default": []
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "isActive",
          "isSuperuser",
          "isVerified",
          "username",
          "name",
          "affiliation",
          "createdAt"
        ],
        "title": "UserRead",
        "description": "A user registered in the system"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tekstuserauth"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  },
...
}